feat: Setup routing to show component `masterblock`
Something is wrong. When the root is opened (`localhost:4200`)
"Show latest MasterBlock" is shown as plain text, but it is supposed to be a link.

When `localhost:4200/masterblock` is typed in the addressbar it directly reverts back to `localhost:4200`.

In the console of my Brave-browser I see the following error:
ERROR NullInjectorError: R3InjectorError(Environment Injector)[_BackendService -> _HttpClient -> _HttpClient]:
  NullInjectorError: No provider for _HttpClient!
    at NullInjector.get (core.mjs:1654:27)
    at R3Injector.get (core.mjs:3093:33)
    at R3Injector.get (core.mjs:3093:33)
    at injectInjectorOnly (core.mjs:1100:40)
    at Module.ɵɵinject (core.mjs:1106:42)
    at Object.BackendService_Factory [as factory]

While in `backend.service.ts` `HttpClient` _is_ imported.